requirements.txt
absl-py==2.1.0
astunparse==1.6.3
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.28
azure-core==1.31.0
certifi==2024.8.30
charset-normalizer==3.3.2
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.10
isodate==0.6.1
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
msrest==0.7.1
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
protobuf==4.25.5
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
setuptools==75.1.0
six==1.16.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
